figurine/familiar starting inventory
     While testing the figurine timer patch, I observed
The goblin wields a crude dagger.
You see a goblin drop out of your pack!

Rather than try to get the sequencing right, just prevent monsters made
by figurine activation or by the create familiar spell start without any
inventory.  This will have a side effect of making wishes for a blessed
figurine of an archon be less powerful, because the subsequent pet A won't
come equipped with a shield of reflection and an artifact--or at least
rustproof--long sword anymore.