latex Guidebook [2 of 2] (trunk only)
     The documentation for symset also changed Guidebook.tex to use the
hyperref package, which the old DECUS TeX distribution I'm using doesn't
have.  I can't remember any discussion about inserting URLs into the
Guidebook and using LaTeX to generate html output.  If there was, no
comments to that effect made it into the .tex file or the cvs log text.
So I'm guessing that \usepackage{hyperref} was a work-around for the
font issue (below) and that the latter was a side-effect of converting
from deprecated \documentstyle{} to recommended \documentclass{}.

     I tried installing hyperref after tracking it down, but using it
generated complaints about several other packages either being too old
or missing entirely.  Coping with them would be too much hassle.  I also
tried just commenting it out, but that results in a font warning that I
assume isn't present when it gets used.  I managed to cobble together
a fix for that, but since hyperref.sty isn't actually needed by our
Guidebook, it was simpler to revert to the way things were done back in
the old days....