seeing a monster make itself invisible
     From the newsgroup:  you could get "suddenly you cannot see the <mon>"
even though it remained visible.  Cited case was for an orc who drank a
potion of invisibility while being observed by a hero wielding Sting, which
causes orcs to be displayed even when they're invisible.  But it could also
happen when non-blind telepathy or extended monster detection is in effect.