bhitpile lint bit
The 'zdir' function parameter has the same name as 'zdir' global
variable, triggering a 'shadowing' warning.  I had to read the 'if'
statement multiple times to convince myself it was doing what was
intended.  It was, but I think this rewrite is easier to understand
(at least for my feeble reptilian brain).

I don't know who Tim Wright is, but his 15 mintues of fame has
lasted for at least a decade so I cut his comment out.  (The 28.5
year old GAN one a dozen lines lower was a tempting target for
removal, but I managed to stop myself; otherwise it never ends.)